Add error support for DepositAddress
