+++ CybridSDK/Components/DepositAddress/Extensions/_Content.swift
@@ -1,79 +1,13 @@
 //
-//  DepositAddresView+Extensions.swift
+//  _Content.swift
 //  CybridSDK
 //
-//  Created by Erick Sanchez on 14/07/23.
+//  Created by Erick Sanchez on 27/09/23.
 //
 
 import UIKit
 
 extension DepositAddresView {
-
-    internal func depositAddressViewLoading() {
-
-        // -- Loading Label Container
-        let loadingLabelContainer = UIView()
-        self.addSubview(loadingLabelContainer)
-        loadingLabelContainer.centerVertical(parent: self)
-        loadingLabelContainer.centerHorizontal(parent: self)
-        loadingLabelContainer.constraint(attribute: .height,
-                                         relatedBy: .equal,
-                                         toItem: nil,
-                                         attribute: .notAnAttribute,
-                                         constant: UIValues.loadingLoadingLabelContainer.height)
-
-        // -- Label
-        let loadingLabel = UILabel()
-        loadingLabel.font = UIValues.loadingAssetNameFont
-        loadingLabel.textColor = UIValues.loadingAssetNameColor
-        loadingLabel.textAlignment = .center
-        loadingLabelContainer.addSubview(loadingLabel)
-        loadingLabel.constraint(attribute: .top,
-                                relatedBy: .equal,
-                                toItem: loadingLabelContainer,
-                                attribute: .top,
-                                constant: 0)
-        loadingLabel.constraint(attribute: .leading,
-                                relatedBy: .equal,
-                                toItem: loadingLabelContainer,
-                                attribute: .leading,
-                                constant: 0)
-        loadingLabel.constraint(attribute: .trailing,
-                                relatedBy: .equal,
-                                toItem: loadingLabelContainer,
-                                attribute: .trailing,
-                                constant: 0)
-        loadingLabel.constraint(attribute: .height,
-                                relatedBy: .equal,
-                                toItem: nil,
-                                attribute: .notAnAttribute,
-                                constant: UIValues.loadingLabelSize.height)
-
-        self.depositAddressViewModel?.loadingLabelUiState.bind { [self] state in
-            switch state {
-            case .VERIFYING:
-                loadingLabel.setLocalizedText(key: UIStrings.loadingLabelVerifying, localizer: localizer)
-            case .GETTING:
-                loadingLabel.setLocalizedText(key: UIStrings.loadingLabelGetting, localizer: localizer)
-            case .CREATING:
-                loadingLabel.setLocalizedText(key: UIStrings.loadingLabelCreating, localizer: localizer)
-            }
-        }
-
-        // -- Spinner
-        let spinner = UIActivityIndicatorView(style: .medium)
-        spinner.color = UIColor.init(hex: "#007AFF")
-        spinner.startAnimating()
-        loadingLabelContainer.addSubview(spinner)
-        spinner.constraint(attribute: .top,
-                           relatedBy: .equal,
-                           toItem: loadingLabel,
-                           attribute: .bottom,
-                           constant: UIValues.loadingSpinnerTopMargin)
-        spinner.centerHorizontal(parent: loadingLabelContainer)
-        spinner.setConstraintsSize(size: UIValues.loadingSpinnerSize)
-        spinner.transform = CGAffineTransform(scaleX: 2.0, y: 2.0)
-    }
 
     internal func depositAddressViewContent() {
 
@@ -350,194 +284,5 @@ extension DepositAddresView {
             createPaymentButton.constraintBottom(contentView, margin: 10)
             createPaymentButton.constraintHeight(45)
         }
-    }
-
-    internal func depositAddressView_Error() {
-
-        // -- Error Container
-        let errorContainer = UIView()
-        self.addSubview(errorContainer)
-        errorContainer.centerVertical(parent: self)
-        errorContainer.constraint(attribute: .leading,
-                                  relatedBy: .equal,
-                                  toItem: self,
-                                  attribute: .leading,
-                                  constant: UIValues.errorContainerHorizontalMargin)
-        errorContainer.constraint(attribute: .trailing,
-                                  relatedBy: .equal,
-                                  toItem: self,
-                                  attribute: .trailing,
-                                  constant: -UIValues.errorContainerHorizontalMargin)
-        errorContainer.constraint(attribute: .height,
-                                  relatedBy: .equal,
-                                  toItem: nil,
-                                  attribute: .notAnAttribute,
-                                  constant: UIValues.errorContainerHeight)
-
-        // -- Icon
-        let icon = UIImageView(image: UIImage(
-            named: "kyc_error",
-            in: Bundle(for: Self.self),
-            with: nil))
-        errorContainer.addSubview(icon)
-        icon.centerHorizontal(parent: errorContainer)
-        icon.constraint(attribute: .top,
-                        relatedBy: .equal,
-                        toItem: errorContainer,
-                        attribute: .top,
-                        constant: UIValues.errorContainerIconTopMargin)
-        icon.setConstraintsSize(size: UIValues.errorContainerIconSize)
-
-        // -- Title
-        let title = UILabel()
-        title.font = UIValues.contentDepositAddressSubTitleFont
-        title.textColor = UIColor.black
-        title.textAlignment = .center
-        title.setLocalizedText(key: UIStrings.error, localizer: localizer)
-        errorContainer.addSubview(title)
-        title.constraint(attribute: .top,
-                         relatedBy: .equal,
-                         toItem: icon,
-                         attribute: .bottom,
-                         constant: UIValues.errorContainerTitleMargins.top)
-        title.constraint(attribute: .leading,
-                         relatedBy: .equal,
-                         toItem: errorContainer,
-                         attribute: .leading,
-                         constant: UIValues.errorContainerTitleMargins.left)
-        title.constraint(attribute: .trailing,
-                         relatedBy: .equal,
-                         toItem: errorContainer,
-                         attribute: .trailing,
-                         constant: -UIValues.errorContainerTitleMargins.right)
-        title.constraint(attribute: .height,
-                         relatedBy: .equal,
-                         toItem: nil,
-                         attribute: .notAnAttribute,
-                         constant: 50)
-
-        // -- Button
-        let done = CYBButton(title: localizer.localize(with: UIStrings.errorButton), action: {
-
-            if let parentViewController = self.superview?.next as? UIViewController {
-                if parentViewController.navigationController != nil {
-                    parentViewController.navigationController?.popViewController(animated: true)
-                } else {
-                    parentViewController.dismiss(animated: true)
-                }
-            }
-        })
-        self.addSubview(done)
-        done.constraint(attribute: .leading,
-                        relatedBy: .equal,
-                        toItem: self,
-                        attribute: .leading,
-                        constant: UIValues.errorContainerButtonMargins.left)
-        done.constraint(attribute: .trailing,
-                        relatedBy: .equal,
-                        toItem: self,
-                        attribute: .trailing,
-                        constant: -UIValues.errorContainerButtonMargins.right)
-        done.constraint(attribute: .bottom,
-                        relatedBy: .equal,
-                        toItem: self,
-                        attribute: .bottom,
-                        constant: -UIValues.errorContainerButtonMargins.bottom)
-        done.constraint(attribute: .height,
-                        relatedBy: .equal,
-                        toItem: nil,
-                        attribute: .notAnAttribute,
-                        constant: 48)
-        done.isHidden = true
-    }
-
-    // -- MARK: Helpers
-    internal func setupScrollView() -> UIView {
-
-        let contentView = UIView()
-        let scrollView = UIScrollView()
-        self.addSubview(scrollView)
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            scrollView.topAnchor.constraint(equalTo: topAnchor),
-            scrollView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            scrollView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            scrollView.trailingAnchor.constraint(equalTo: trailingAnchor)
-        ])
-
-        scrollView.addSubview(contentView)
-        contentView.translatesAutoresizingMaskIntoConstraints = false
-        let heightConstraint = contentView.heightAnchor.constraint(equalTo: scrollView.heightAnchor)
-        heightConstraint.priority = UILayoutPriority(250)
-        NSLayoutConstraint.activate([
-            contentView.topAnchor.constraint(equalTo: scrollView.topAnchor),
-            contentView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-            contentView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
-            contentView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-            contentView.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
-            heightConstraint
-        ])
-        return contentView
-    }
-
-    internal func hideView(_ view: UIView) {
-
-        view.removeConstraint(attribute: .height)
-        view.constraintHeight(0)
-        view.removeConstraint(attribute: .top)
-    }
-
-    @objc func copyDepositAddressToClipboard(_ sender: UITapGestureRecognizer) {
-        let address = self.depositAddressViewModel?.currentDepositAddress?.address ?? ""
-        UIPasteboard.general.string = address
-    }
-}
-
-extension DepositAddresView {
-
-    enum UIValues {
-
-        // -- Size
-        static let loadingLoadingLabelContainer = CGSize(width: 0, height: 80)
-        static let loadingLabelSize = CGSize(width: 0, height: 28)
-        static let loadingSpinnerSize = CGSize(width: 43, height: 43)
-        static let errorContainerHeight: CGFloat = 100
-        static let errorContainerIconSize = CGSize(width: 40, height: 40)
-
-        // -- Margin
-        static let loadingAssetContainerTopMargin: CGFloat = 30
-        static let loadingAssetNameTopMargin: CGFloat = 12
-        static let loadingSpinnerTopMargin: CGFloat = 15
-        static let errorContainerHorizontalMargin: CGFloat = 20
-        static let errorContainerIconTopMargin: CGFloat = 15
-        static let errorContainerTitleMargins = UIEdgeInsets(top: 5, left: 10, bottom: 0, right: 10)
-        static let errorContainerButtonMargins = UIEdgeInsets(top: 15, left: 15, bottom: 0, right: 15)
-
-        // -- Font
-        static let loadingAssetNameFont = UIFont.make(ofSize: 24)
-        static let contentDepositAddressSubTitleFont = UIFont.make(ofSize: 13)
-
-        // -- Color
-        static let loadingAssetNameColor = UIColor.init(hex: "#3A3A3C")
-        static let contentDepositAddressWarningColor = UIColor.init(hex: "#636366")
-    }
-
-    enum UIStrings {
-
-        static let loadingLabelVerifying = "cybrid.deposit.address.loading.label.verifying"
-        static let loadingLabelGetting = "cybrid.deposit.address.loading.label.getting"
-        static let loadingLabelCreating = "cybrid.deposit.address.loading.label.creating"
-        static let contentDepositAddressTitle = "cybrid.deposit.address.content.deposit.address.title"
-        static let contentDepositAddressQRWarningOne = "cybrid.deposit.address.content.qr.warning.1"
-        static let contentDepositAddressQRWarningTwo = "cybrid.deposit.address.content.qr.warning.2"
-        static let contentDepositAddressNetworkTitle = "cybrid.deposit.address.content.deposit.network.title"
-        static let contentDepositAddressAddressTitle = "cybrid.deposit.address.content.deposit.address.address.title"
-        static let contentDepositAddressTagTitle = "cybrid.deposit.address.content.deposit.tag.title"
-        static let contentDepositAddressAmountTitle = "cybrid.deposit.address.content.deposit.amount.title"
-        static let contentDepositAddressMessageTitle = "cybrid.deposit.address.content.deposit.message.title"
-        static let contentWarning = "cybrid.deposit.address.content.warning"
-        static let contentDepositAddressButton = "cybrid.deposit.address.content.button"
-        static let error = "cybrid.deposit.address.error"
-        static let errorButton = "cybrid.deposit.address.error.button"
     }
 }
