+++ CybridSDK/Core/ComponentEnum.swift
@@ -10,4 +10,5 @@ import Foundation
 public enum ComponentEnum {
 
     case walletsComponent
+    case depositAddressComponent
 }
