+++ CybridSDK/Utilities/Localization/fr.lproj/Localizable.strings
@@ -217,3 +217,4 @@
 "cybrid.server.error.generic" = "Erreur de serveur, réessayez.";
 "cybrid.server.error.wallets.view.error.invalid_parameter.tag" = "La balise ne doit pas être vide";
 "cybrid.server.error.wallets.view.error.data_exists" = "L'adresse existe déjà.";
+"cybrid.server.error.deposit.address.create.error.unverified_customer" = "Le client n'a pas été vérifié";
